backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-file-processor"
version = "1.0.0"
description = "AI-powered file processing backend with FastAPI"
requires-python = ">=3.12"
authors = [{name = "AI File Processor"}]
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "openai>=1.54.0",
    "anthropic>=0.39.0",
    "python-magic>=0.4.27",
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]  # Line length handled by formatter

[tool.setuptools]
packages = ["app"]
